Make sure OmniAuth is hit with a direct post
Omniauth requires a POST, turbo is using fetch() which results in a failed 404 call. 
Because we'll be redirected anyway, I don't see any reason to use Turbo.
Tested locally, will see if there are any specs around this. 

Also, I'm assuming that this file somehow serves as the erb template when you are using the super_scaffolder?